move plan type constants to registry fix typing & imports
