+++ src/artemis/experiment_plans/__init__.py
@@ -1,5 +1,1 @@
 """This module contains the experimental plans which artemis can run."""
-
-from artemis.experiment_plans.experiment_registry import PLAN_REGISTRY
-
-__all__ = ["PLAN_REGISTRY"]
