+++ src/artemis/external_interaction/callbacks/fgs/ispyb_callback.py
@@ -1,5 +1,7 @@
+from __future__ import annotations
+
 import os
-from typing import Dict
+from typing import TYPE_CHECKING, Dict
 
 from bluesky.callbacks import CallbackBase
 
@@ -10,7 +12,9 @@ from artemis.external_interaction.ispyb.store_in_ispyb import (
 )
 from artemis.log import LOGGER
 from artemis.parameters.constants import ISPYB_PLAN_NAME, SIM_ISPYB_CONFIG
-from artemis.parameters.internal_parameters import InternalParameters
+
+if TYPE_CHECKING:
+    from artemis.parameters.internal_parameters import InternalParameters
 
 
 class FGSISPyBHandlerCallback(CallbackBase):
