+++ src/artemis/external_interaction/callbacks/fgs/nexus_callback.py
@@ -1,4 +1,6 @@
-from typing import Optional
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Optional
 
 from bluesky.callbacks import CallbackBase
 
@@ -8,7 +10,9 @@ from artemis.external_interaction.nexus.write_nexus import (
     create_parameters_for_second_file,
 )
 from artemis.log import LOGGER
-from artemis.parameters.internal_parameters import InternalParameters
+
+if TYPE_CHECKING:
+    from artemis.parameters.internal_parameters import InternalParameters
 
 
 class FGSNexusFileHandlerCallback(CallbackBase):
