+++ src/artemis/external_interaction/system_tests/conftest.py
@@ -7,11 +7,11 @@ from ispyb.sqlalchemy import DataCollection
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
+import artemis.parameters.internal_parameters as ip
 from artemis.external_interaction.ispyb.store_in_ispyb import (
     StoreInIspyb2D,
     StoreInIspyb3D,
 )
-from artemis.parameters.internal_parameters import InternalParameters
 from artemis.utils import Point3D
 
 ISPYB_CONFIG = "/dls_sw/dasc/mariadb/credentials/ispyb-dev.cfg"
@@ -42,7 +42,7 @@ def fetch_comment() -> Callable:
 
 @pytest.fixture
 def dummy_params():
-    dummy_params = InternalParameters()
+    dummy_params = ip.InternalParameters()
     dummy_params.artemis_params.ispyb_params.upper_left = Point3D(100, 100, 50)
     dummy_params.artemis_params.ispyb_params.pixels_per_micron_x = 0.8
     dummy_params.artemis_params.ispyb_params.pixels_per_micron_y = 0.8
