+++ src/artemis/parameters/constants.py
@@ -1,8 +1,4 @@
 from enum import Enum
-from typing import Union
-
-from artemis.devices.fast_grid_scan import GridScanParams
-from artemis.devices.rotation_scan import RotationScanParams
 
 SIM_BEAMLINE = "BL03S"
 SIM_INSERTION_PREFIX = "SR03S"
@@ -12,13 +8,6 @@ DEFAULT_EXPERIMENT_TYPE = "grid_scan"
 
 PARAMETER_VERSION = 0.1
 
-EXPERIMENT_DICT = {
-    "fast_grid_scan": GridScanParams,
-    "rotation_scan": RotationScanParams,
-}
-EXPERIMENT_NAMES = list(EXPERIMENT_DICT.keys())
-EXPERIMENT_TYPE_LIST = list(EXPERIMENT_DICT.values())
-EXPERIMENT_TYPES = Union[GridScanParams, RotationScanParams]
 SIM_ISPYB_CONFIG = "src/artemis/external_interaction/unit_tests/test_config.cfg"
